Placement Engine: Abstract methods into interfaces Task and Host
Summary:
This diff sets up the code so that we will be able to replace the function signatures
of the placement.Strategy interface. Here we move some more of the v0 specific things
into the `models` package, and define those interfaces.
In the next diff, we will add the `HostFilter` specific methods to those interfaces,
as well as change the function signatures as mentioned above.

Reviewers: adityacb, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2928519