Fix issue of RMTask with older run waiting to be launched
Summary:
Currently, it is possible for an older run of a task (that has been admitted) to be stuck (in RM) waiting for launch.
Consider the below scenario
1. JM polls RM to get placements (the RMTasks corresponding to the tasks in placements transit to LAUNCHING)
2. Before processing the placements, JM fails over.
3. New JM leader receives update request for task(s) and tries to enqueue the tasks (EnqueueGangs) which fails since the tasks are already in the tracker (in LAUNCHING state)

In the above scenario, the tasks (older run) are stuck in RM waiting for their launch and JM keeps (re)trying to enqueue the new run of the tasks.

This diff addresses the above issue by skipping the older run of the task (which triggers the killing of the corresponding RMTask).

Test Plan: - Added unit test

Reviewers: apoorvaj, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Tags: #peloton

Maniphest Tasks: T3106625

Differential Revision: https://code.uberinternal.com/D2939969