Placement Engine: More abstraction for plugins libs
Summary:
Work in progress, plugins library is now completely separate, and
no longer knows anything about:
- v0api.hostsvc.HostFilter
    - Includes RankHint, PlacementStrategy, HostHint
- v0api.resmgr.Task
    - Includes every sub field
- v0api.hostsvc.Offer
    - Includes every sub field

We redefine `plugins.HostFilter` to be `plugins.HostDemands` to represent something that
can be turned into a HostFilter when it goes out to the HM.
We add the required methods on Assignment to fit the Task interface (that was also changed a bit).

Also renamed the `Filters` method to `GroupTasksByPlacementNeeds`, to better describe what the
method is meant to do.

NOTE: We lose type safety on RankHint and Constraint. This is a trade off we are making to
have the rest be cleaned up.

Reviewers: O521 Project Peloton: Add reviewers, adityacb

Reviewed By: O521 Project Peloton: Add reviewers, adityacb

Subscribers: adityacb, jenkins

Differential Revision: https://code.uberinternal.com/D2948665