[DB schema + code] Start writing JobSpec blob to DB as a new job_config column
Summary:
For P2K, we are relying on `stateless.JobSpec` and `pod.PodSpec` directly
to launch pods. To that end, we need to store the original `JobSpec` as a blob
along with the converted `JobConfig` in DB.

This diff changes the `cached/job.go` code as well as the `JobConfigOps` to
support writing JobSpec to the DB.

To retrieve JobSpec, I thought it was better to return a new result struct
`JobConfigOpsResult` that contains all the columns unmarshalled into their
respective structs, instead of adding a specialized `GetJobSpec` method.

There are a few function signature changes in job.go for:
- `cachedJob.Update()`
- `cachedJob.CompareAndSetConfig()`
- `cachedJob.Create()`
- `cachedJob.RollingCreate()`

This results in a lot of test code and call site modifications to add `JobSpec`
as a parameter. They key point to note here is that the only valid places
where a valid `JobSpec` will be sent in these calls is from the stateless API handlers.
Rest of the callsites will just pass in a `nil`.

Most of the code changed in this diff is in:
- `pkg/jobmgr/cached/job.go`
- `pkg/storage/objects/job_config.go`

Reviewers: O521 Project Peloton: Add reviewers, amitbose, apoorvaj, zhixin, #p2k

Reviewed By: O521 Project Peloton: Add reviewers, zhixin

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2975235