Notify listeners when task runtime is replaced in cache
Summary:
Today, we notify listeners only when the task runtime changes are successfully written to DB. However, it is possible that the DB write may time out (even though the write went through) and we invalidate the cache. When the task is reloaded into cache, the listeners are not notified causing them to miss this task event.

This diff addresses the above issue by notifying the listeners whenever the task cache is updated.

Test Plan: Unit tests

Reviewers: apoorvaj, zhixin, O521 Project Peloton: Add reviewers, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, amitbose

Subscribers: amitbose, jenkins

Differential Revision: https://code.uberinternal.com/D2974605