[DB schema and code] Add podspec and version to task_config table
Summary:
Added spec and api_version to the task_config.
Added code to `CreateTaskConfig` to accept `pbpod.PodSpec`.

Not adding task_config to ORM as part of this diff.
That change will be worked on in next couple of weeks.

Added a call to `GetPodSpec` to fetch only the podSpec.
Next diff will modify the `launcher.go` to read the podSpec and
call `hostsvc.LaunchPods` API if jobmgr is configured for k8s

Reviewers: O521 Project Peloton: Add reviewers, sishi, zhixin, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, amitbose

Subscribers: amitbose, sishi, jenkins

Differential Revision: https://code.uberinternal.com/D2993159