Add PARALLEL_STATELESS_UPDATE and STATELESS_CREATE tables in performance tests
Summary:
In the current Peloton performance reports, two tables, PARALLEL_STATELESS_UPDATE and STATELESS_CREATE, are missing. This diff is to format from the generated dataframe, and add the tables to the performance reports.

It supposes to look like:

[PARALLEL_STATELESS_UPDATE] table
   NumJobs  NumStartTasks  Sleep(s)  BatchSize  AverageTimeInSeconds(Baseline) AverageTimeInSeconds(Current) Execution Duration Change (%)
0      100             90      1000         10             94.944835 .... ..... ......

[STATELESS_CREATE] table
   NumStartTasks  Sleep(s)   TotalTimeInSeconds(Baseline) TotalTimeInSeconds(Current) Execution Duration Change (%)
0           9000      1000            81.70486 .... .... ....

Test Plan: Jenkins configuration currently doesn't allow us to efficiently test locally. We plan to push to production, and check the generated performance reports to verify if further changes are needed.

Reviewers: O521 Project Peloton: Add reviewers, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2979257