Host Manager: Added TerminateLeases protobuf definition
Summary:
Added the protobuf definition for the HostManager `TerminateLeases`
internal RPC call.

Ideally, `TerminateLeases` would only take in a list of `LeaseID`. However,
the `HostCache` internals only contain an index of the hosts on their hostnames,
so we need to take in both `LeaseID` and `hostname` for each lease we want
to terminate.
At this time, we would like to avoid maintaining a new index of the hosts
on their `LeaseID`, because that would lead to more complex code in the
`HostCache` and more tricky synchronization that is prone to errors.

Reviewers: adityacb, yiran, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2996359