ORM: Introduce type `OptionalString` for string primary keys
Summary:
If a type `MyObject` is defined as a struct with a `string` field, initializing a such empty object as `&MyObject{}` results in a string field set to the zero value of string type: the empty string `""`. The ORM `GetAll()` function is called with an empty object such as:  `ormClient.GetAll(ctx, &MyObject{})`. If in ORM `MyObject` is defined with its PimaryKey (PK) of type string (but without any PK value set since there is no PK selection intended), the storage code builds a `SELECT` query based on the value of the PK and so builds a `WHERE` clause such as `PK.name = ""`.

To avoid this, it is necessary to separate the case between a PK of type string having an actual non-zero value, and when this PK of type string is just  initialized and set to its zero value `""`.

This can be solved by introducing a custom type `OptionalString` which can either represent a valid string value, or `nil` in case unset to a proper string value.

In Cassandra the PK field remains of type `text`, the ORM layer has the responsibility of converting the string/text field to/from the custom type `OptionalString` used in the ORM object.

Test Plan:
- Updated unit tests
- Integration tested with `HostInfo` ORM object, most notably the `GetAll()`

Reviewers: O521 Project Peloton: Add reviewers, adityacb, amitbose, sishi

Reviewed By: O521 Project Peloton: Add reviewers, amitbose

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2962075