Fix error handling in store.deletePodEventsOnDeleteJob
Summary: In `GetCurrentVersion`, we are wrapping the original error with an error message for help in tracking. In store.deletePodEventsOnDeleteJob, while handling error, we need to look at the original(underlying) error rather the topmost error (the error returned by `GetCurrentVersion`) , to determine the yarpc error code. This diff addresses the above issue

Reviewers: adityacb, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: zhixin, jenkins

Differential Revision: https://code.uberinternal.com/D3017725