Fix desired host placement deadline
Summary:
In placement engine, there are two deadlines.
One deadline is the deadline to place a task of a
particular type on a qualified host.
The other deadline is the deadline to place a task
on the desired host, which is used for in-place update.
The second deadline should not be subject to the first.
For example, if a normal stateless task can wait
placement for 60s, but in-place update task can wait for
120s, then the task should be subject to the 120s limit
instead of 60s limit. This diff fixes the overwrite of
deadline.

Reviewers: varung, adityacb, pourchet, O521 Project Peloton: Add reviewers

Reviewed By: pourchet

Subscribers: jenkins

Maniphest Tasks: T3304323

Differential Revision: https://code.uberinternal.com/D3016897