Handle placement with stale run-id in resmgr tracker
Summary:
Consider the below scenario.
1. Task (run_id=1) is in resmgr in PLACING state.
2. Task (run_id=1) in resmgr is killed by jobmgr (say due to launch error).
3. Jobmgr enqueues task (run_id=2) into resmgr.
4. Task (run_id=2) gets admitted, and goes to PLACING state.
5. Placement Engine writes placement for task (run_id=1) to resmgr.

In the above scenario, currently, resmgr checks the state of the RMTask (run_id=2) in tracker to ensure it is in the right state but doesn't verify the run-id of the task in placement matches to that in tracker. This causes resmgr to misinterpret the placement is for the task in the tracker (run_id=2)

This diff addresses the above issue by checking the run-id of the task in the placement with that of that task in the tracker.

Test Plan:
 - Tested locally on minicluster
 - Enable `test__in_place_update_host_maintenance`

Reviewers: varung, avyas, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Tags: #peloton

Maniphest Tasks: T3290279

Differential Revision: https://code.uberinternal.com/D2996825