Host Maintenance: add host state validation in handler
Summary: Handlers should as a first step validate the request, in case the request is invalid then avoiding any further processing of the request by returning a YARPC error right away. The `StartMaintenance()` handler should validate the requested hostname is registered as an UP agent, avoiding contacting MesosMaster in case the hostname is invalid.

Reviewers: O521 Project Peloton: Add reviewers, sachins

Reviewed By: O521 Project Peloton: Add reviewers, sachins

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3008053