[Part 1] Launch pods from Jobmgr
Summary:
This diff lays the ground work to add some config flags in jobmgr to
launch pods vs tasks based on if it is a k8s cluster.

I tried really hard to do this cleanly, by moving a lot of functionality from
`launcher` to `placement` but failed because some of the stateful code
directly accesses launcher API from goalstate. We can refactor this after
deleting stateful legacy code.

To unblock P2K, this diff contains a few if checks based on k8s flag in the launcher.

I verified that for a stateless job, with k8s configured, we are actually
able to acquire hosts via placement engine using the v1alpha hostmgr API
but immediately return those offers back to hostmgr.

For now, I plan to land this diff so someone can debug whats going on in the
v1 offer service of placement engine.

Reviewers: O521 Project Peloton: Add reviewers, #p2k, pourchet, evelynl, binz

Reviewed By: #p2k, pourchet, binz

Subscribers: binz, jenkins

Differential Revision: https://code.uberinternal.com/D3047569