Make PREEMPTING a timeout state in rmtask
Summary:
Consider the following scenario

1. Resmgr selects task t (RUNNING) for preemption and enqueues into PreemptionQueue
2. Jobmgr calls GetPreemptibleTasks to get tasks to preempt - task t transits to PREEMPTING
3. Jobmgr crashes before killing task t.
In the above scenario, task t is stuck in PREEMPTING state as we never enqueue a PREEMPTING task into the preemption queue.
This diff adds a timeout to rmtask PREEMPTING state so that the task t transits to RUNNING on timeout.

Test Plan: Added unit tests

Reviewers: avyas, O521 Project Peloton: Add reviewers

Reviewed By: avyas, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3075823