Skip default task config not found error due task_config deprecation
Summary:
In some early testing cluster, we are seeing a crash in jobmgr due
to failure to recover task config for a job.

After inspection, it seems like the task config for that particular job
was a test job created around Sep 2018, and its task config still exists
in task_config table.

Skip crashing on default task config not found error in jobmgr / resmgr
recovery code.

Reviewers: adityacb, zhixin, avyas, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins, amitbose

Differential Revision: https://code.uberinternal.com/D3082663