Integration Test: Increase stateless respool capacity to cluster capacity
Summary:
Today, stateless job run in a respool which is about 50% of the cluster capacity. This doesn't allow us to test scenarios where the cluster is operating near it's full capacity (e.g. pausing of host draining when SLA of job(s) is violated due to lack of resources in the cluster)

This diff does the following
1. Create a respool for stateless jobs with capacity equal to cluster capacity
2. Remove `test_job`(`long_running_job`+`stateless_job`) fixture, replacing the tests using the fixture with `long_running_job` fixture and duplicate the tests into stateless job tests.

Reviewers: zhixin, amitbose, O521 Project Peloton: Add reviewers, varung

Reviewed By: amitbose, O521 Project Peloton: Add reviewers, varung

Subscribers: varung, jenkins

Differential Revision: https://code.uberinternal.com/D3090615