Use LOCAL_SERIAL for CAS operations on Cassandra
Summary:
Peloton has two rings of Cassandra cluster across physical regions, and in a staging cluster we hit a scenario where one of the ring was unhealthy which lead to failure of operations using CAS. On further investigation, we found that it was using SERIAL consistency for CAS operations which need cross region concensus.

To mitigate the issue and aligning with Peloton to be zone local. Changing CAS consistency level to LOCAL_SERIAL.

Reviewers: kevinxu, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: kevinxu, O521 Project Peloton: Add reviewers

Subscribers: adityacb, jenkins

Maniphest Tasks: T3511635

Differential Revision: https://code.uberinternal.com/D3096593