Handle 'instancesToBeRetried' in cachedjob.PatchTasks caller
Summary:
One of the values returned by PatchTasks is `instancesToBeRetried` which are the instances that are not present in cache. It is the responsibility of the caller to retry the action after reloading the corresponding task runtimes.

This diff does the following
1. In PatchTasks, request task runtime to be reloaded only if necessary. E.g. if the instance is already marked unavailable, we can let the patch go through since it can't violate the SLA.
2. Handle `instancesToBeRetried` in every PatchTasks call

Test Plan: Added unit tests

Reviewers: O521 Project Peloton: Add reviewers, zhixin

Reviewed By: O521 Project Peloton: Add reviewers, zhixin

Subscribers: jenkins

Maniphest Tasks: T3376899

Differential Revision: https://code.uberinternal.com/D3081195