Get taskConfig from legacy task_config table if needed
Summary:
There are still a number of legacy jobs that have config in
task_config table and NOT task_config_v2 table. For such jobs,
my initial hope was to not recover them and let them be killed
as orphan tasks. That assumption seems invalid as they seem
to get stuck in the goalstate engine on events.

To avoid this, I am bringing back code to read from legacy
task_config table but also back fill the legacy configs into
the new task_config_v2 table.

Reviewers: O521 Project Peloton: Add reviewers, zhixin

Reviewed By: O521 Project Peloton: Add reviewers, zhixin

Subscribers: zhixin, jenkins

Differential Revision: https://code.uberinternal.com/D3108395