Fix flaky test__kill_mesos_agent_makes_task_resume
Summary:
The batch suite integration test `test__kill_mesos_agent_makes_task_resume` was failing since the agents were not coming back up soon enough.
As a result, TASK_LOST event was sent out for the tasks. With no retry policy configured for the job, the job transitioned to FAILED state.

This diff fixes the test by specifying a retry policy to the job so that job is retried after TASK_LOST is received.

Reviewers: amitbose, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3107909