Update unavailable instances before updating available ones
Summary:
Today, for each update batch, we do the the following
1. Add new instances
2. Update existing instances
3. Remove instances

For 2, we try to update instances in the order of their instanceID which can result in updating an
available (RUNNING and HEALTHY) instance before an unavailable one thereby increasing the number of
unavailable instances. This can cause the update workflow to get stuck (although there is a way
for it to progress)

This diff orders the instances to be updated in the order following order
1. unavailable-instances
2. killed-instances
3. instances with unknown availability
4. available-instances

This will ensure we keep the number of unavailable instances to
a minimum thereby giving the update workflow the best chance to progress.

Test Plan:
 - Unit tests
 - Tested on local minicluster
      1. Create a job, manually kill some the tasks, update the job with batch_size=1, and verify that the killed tasks are updated first.
      2. Create a job, some task with bad config, wait for a while so task get throttled, update the job, and verify that failed tasks get updated first.

Reviewers: zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3098755