Implement host-pool change and publish event
Summary:
Add code to change the host-pool associated with a host.
Also generate a event on the host-manager evenstream when
this happens. A new event-type is defined for this
because existing events are task specific. The new event
is defined only in v0 at this time; v1alpha will come
in a subsequent diff. Note that because of this new
event type, guards had to be added in a bunch of files
such as common/statusupdate, jobmgr/task/event,
resmgr/handler and hostmgr/handler. The code in these
files was written with the assumption that events are
for task status updates only.

Test Plan: Unit-tests

Reviewers: O521 Project Peloton: Add reviewers, yunpeng

Reviewed By: yunpeng

Subscribers: apoorvaj, jenkins

Maniphest Tasks: T3554395

Differential Revision: https://code.uberinternal.com/D3133171