Fix resource cleanup for older runs of the task
Summary:
Before, if the terminal mesos event for an older run arrived after the latest run of the task has been cleaned up from the tracker, then the resources were never released due to a bug in `tracker.MarkItDone`. We were checking if the task exists in tracker before cleaning up.  As a result, if a terminal event for an older run arrives after the newer run has been cleaned up from tracker, it would error out and not release the resources held-up by the older run.

This diff addresses the above issue by not checking for the task in the tracker before cleaning up the task.

This diff also modifies `MarkItDone` and `MarkItInvalid` to take only `mesos-task-id` as argument (instead of both `peloton-task-id` and `mesos-task-id`).

Test Plan:
Added unit tests

Reviewers: avyas, O521 Project Peloton: Add reviewers

Reviewed By: avyas, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3607291

Differential Revision: https://code.uberinternal.com/D3138239