Integration Tests: First K8S integration test
Summary:
This commit adds a no-op test to the stateless integration tests. This
test will only run when running the k8s integration tests.
This diff also contains a fix to the kind library to make it ignore
certain failures that were encountered locally linked to possibly an old
python version (Peloton is still on 2.7).

For instance we sometimes see a "Bad File Descriptor" exception when tearing down the k8s
cluster at the end of tests. However, this isn't a problem because we first tear down the cluster before
bootstrapping it anyway at the start of tests.

Reviewers: adityacb, yiran, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3143881