Fix flaky stateless integration tests
Summary:
Some stateless integration tests testing SLA Aware Host Maintenance create a stateless job with host-limit 1 constraint and more than 3 instances. The tests start host maintenance on one of the hosts without waiting for all 3 instances to reach RUNNING state. If the host maintenance is started on a host on which the task is placed(PLACED/LAUNCHED/STARTING) but not yet in healthy, the instance kill will go through (since it doesnt violate SLA - instance is already unavailable), making the tests flaky.

The tests need to be modified to wait for 3 instances to reach RUNNING state before starting maintenance on one of the hosts.

Reviewers: apoorvaj, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3614391

Differential Revision: https://code.uberinternal.com/D3144767