Performance Test: Enable parallel-stateless-update test
Summary:
Before, this test was disabled since Peloton/Mesos would go into a bad state on running this test. The udpate would get stuck with Peloton not receiving offers from a large number of agents (>60% of the cluster capacity). A Mesos Master leader failover was required to fix the issue.

After bumping up the Mesos version to 1.7.1, the test seems to be running fine. Hence, enabling the test

Test Plan: Tested on a vcluster

Reviewers: varung, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2639217

Differential Revision: https://code.uberinternal.com/D3149263