Add hold functionalities in hostsummary and hostcache
Summary:
The same design of in-place upgrade in mesos. I kept similar interface for the convience of future merging work.

The implementation is also similar to what is in offerpool and summary for mesos but I removed the HeldHost state, because:
1. Simplicity. `len(heldPodIDs)` gives you the same information. And also no more `getResetStatus()`.
2. The defition of Held is vague. A host can hold pods in Ready, Placing and Held state.

Also realized that hostsummary should be moved out of hostcache otherwise there is a circular dependency if hostsummary mock is used to test hostcache, but since there is a bigger refactor later, I would not change it now.

Followup: add pruners and add handlers in hostmgr.

Test Plan: unit tests

Reviewers: #p2k, zhixin, adityacb, O521 Project Peloton: Add reviewers

Reviewed By: #p2k, zhixin, adityacb, O521 Project Peloton: Add reviewers

Subscribers: varung, jenkins

Differential Revision: https://code.uberinternal.com/D3156769