Limit number of pod runs to return inside GetTasksWithoutConfigs
Summary:
Limit number of pod runs to return inside GetTasksWithoutConfigs endpoint,
bounded by following conditions:

1. GetTasksPodMax (config) - maximum number of pods (including previous runs)
   to return inside GetTasksWithoutConfigs endpoint.
2. minPodRunsDepth (const) - minimum number of pod runs depth to have (2) if
   PodRunsDepth is configured to be greater than minimum.

This somewhat simulates Aurora's behavior for retraining previous task runs,
although the difference is that Aurora performs periodical pruning of old tasks
instead of setting a hard limit on the read endpoint.

By giving it a upper bound in terms of number of pods to retrieve, it limits the
number of queries made to Peloton, which reduces overall endpoint latency.

However, a minimum pod run depth is enforced so that the downstream service
could use it to display error message when tasks launch failure occurred.

Reviewers: varung, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3679237

Differential Revision: https://code.uberinternal.com/D3156749