Fix sla_violated_jobs metrics in job factory
Summary:
The batch size of a job update can be more than the MaximumUnavailableInstances configured in the job spec. As a result, it is possible that `len(unavailable_instances)>job_spec.sla.maximum_unavailable_instances` in a job with an ongoing update.

This diff skips the SLA violation check for jobs with an ongoing update.

Test Plan: Unit tests

Reviewers: zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3177947