Dedupe job update event for update.writeWorkflowEvents
Summary:
update.writeWorkflowEvents does not check if an update state
changes. It is used by update.Modify which is called in
UpdateStart when the update is in INITIALIZED state.
As a result, user would see two update INITIALIZED state
in job update events. This diff fixes the issue.

Reviewers: sachins, O521 Project Peloton: Add reviewers, apoorvaj

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T3640437

Differential Revision: https://code.uberinternal.com/D3178613