Add Limit support to ORM
Summary:
Added `limit` support to the select query template.
Now, `Get` will by default result in `Limit 1` getting
added to the select query.

This is needed to optimize the way GetPodEvents is
implemented. In that, if a run_id is not supplied, we do
a Get() call instead of a GetAll() to get the max run_id
and then do a `GetAll` call to get all updates for this
run_id.

Earlier, this query did two `GetAll` calls, which is
expensive.

Reviewers: O521 Project Peloton: Add reviewers, sishi, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, sishi

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3198931