Simple go-routines for GetJobs API
Summary:
Simple parallelism to get job config, job update and job update events in parallel for GetJob API.

On profiling for a production cluster, 0.52s spend in GetJob below is the distribution.
```
0.10s - GetJobRuntime
0.22s - GetJobConfig
0.10s - GetUpdate
0.06s - GetJobUpdateEvents
```

With parallelizing the call, we can offset the time it takes for GetUpdate and GetJobUpdateEvents while it is doing GetJobConfig. Overall time for GetJob should reduce from 0.52s to 0.32s.

Test Plan: unit test

Reviewers: apoorvaj, zhixin, sachins, O521 Project Peloton: Add reviewers

Reviewed By: sachins, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3201157