Refactor use of UpdateRecord in GetUpdatesForJob in store.go
Summary:
GetUpdatesForJob returns a list of update_ids for a provided job. It deserialized the return result set in UpdateRecord struct which has 20 fields. STAPI layer uses reflection on fields for UpdateRecord which is unnecessary since we only three fields are returned from Updates View. Adding a new UpdateViewRecord which has three fields as Updates MV.

CPU Profiling, reflects 21% time is spent in FillObject for Updates of which most of the time is in reflection calls on UpdateRecord.
{F151015519}

Reviewers: apoorvaj, zhixin, sachins, O521 Project Peloton: Add reviewers

Reviewed By: sachins, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3203283