Do not start maintenance on a non-Peloton agent
Summary: Before, when host maintenance was triggered via `StartMaintenance` API, Peloton would blindly trigger Mesos Maintenance on the host without checking if the host is registered as Peloton agent. This diff adds the check to see if the host is a Peloton agent before starting maintenance on the host.

Test Plan: Unit tests

Reviewers: amitbose, varung, O521 Project Peloton: Add reviewers

Reviewed By: amitbose, varung, O521 Project Peloton: Add reviewers

Subscribers: apoorvaj, jenkins

Differential Revision: https://code.uberinternal.com/D3209161