Use a valid job spec on rollback path for first deployment
Summary:
In aurora bridge manual rollback path, the current logic is to
get current workflow status, under which, get previous job
config version, use the previous job config version, retrieve
the job spec.

Unfortunately, this does not work for jobs without previous
job configs, i.e. jobs at their very first deployment. In which case,
we will get "0-x-x" entity version back for previous version.
And if we use "0-x-x", we will not get a valid job config.

To workaround this issue, if we see the previous entity version
is "0-x-x", we use current job config instead, but with instance count
changed to 0, so that we can roll back to a state where all tasks are
deleted.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, varung

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj, varung

Subscribers: zhixin, jenkins

Differential Revision: https://code.uberinternal.com/D3221341