Fix task config compare after migrating to use new mesos spec
Summary:
* Hard-code mesos URI "extract" field to be true

Previously when we were using mesos.v1.CommandInfo.URI,
"extract" field was left as unfilled, however, since in
the proto definition the field is default to true, the
actual value we passed to mesos is actually true.

In pod.apachemesos.PodSpec.URI, "extract" field can no
longer have a different value, in order to retain
the previous behavior and to avoid unnecessary instance
restarts, we are hard-code this value to true.

* Use custom comparator for CommandInfo and ContainerInfo

The reason for doing this is that since mesos uses proto2, which allows
default values and distinguishes nil value, but in peloton, we use proto3
and no nil is allowed. So the conversion from peloton's new mesos spec
back to mesos Command / ContainerInfo causes information loss.

Implement a custom comparator to workaround this issue.

Reviewers: apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3819205

Differential Revision: https://code.uberinternal.com/D3228869