Fix kill pod in k8s
Summary:
jobmgr only marks a pod/task as killed when the pod state is KILLED. However, this is not a valid pod phase in k8s.
When pod is deleted, a delete grace period is added to the pod spec and a delete event is sent to all watchers. Then, kubelet, watching the event,
kills the pod and pod is removed from the storage. The final state when pod exits is not reflected in the update.

Also copied over one test that verifies pod kill.

Test Plan:
tested with integration test.

Note: I tried to make test_stateless_job_spec.yaml works for both k8s and mesos but I cant make it work for all tests. Most of the tests actually run with UC, changing to docker breaks ~10 tests.

Reviewers: #p2k, O521 Project Peloton: Add reviewers, yiran, adityacb

Reviewed By: #p2k, O521 Project Peloton: Add reviewers, yiran, adityacb

Subscribers: apoorvaj, adityacb, pourchet, jenkins

Differential Revision: https://code.uberinternal.com/D3226405