Add event id to the v1alpha event stream for dedupe
Summary:
Add EventID to the v1alpha proto for eventstream, so that
we can use this for deduping event stream. For Mesos, this
will be the UUID from the mesos TaskStatusUpdate. For K8S,
this will be the ResourceVersion from the pod status update.

Next up, add southbound dedupe for mesos plugin.
Next next up, add northbound dedupe in v1alpha event stream.

Reviewers: O521 Project Peloton: Add reviewers, evelynl, zhixin

Reviewed By: evelynl

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3241513