Publish host-pool capacity
Summary:
Adding an internal host-mgr API to give the
resource capacity of all host-pools (similar to
cluster capacity). For each host-pool, the capacity
gets cached and it is refreshed periodically with
the reconciliation cycle, or when a host moves
from one pool to another.

Test Plan: Unit-tests

Reviewers: O521 Project Peloton: Add reviewers, yunpeng

Reviewed By: yunpeng

Subscribers: jenkins

Maniphest Tasks: T3574041

Differential Revision: https://code.uberinternal.com/D3263663