P2K: Integ Tests: Allow for code sharing between k8s and mesos tests
Summary:
We used pytest marks to determine what kind of cluster to bring up, which meant that if
a test was k8s AND mesos enabled, we would bring up k8s every time.
Now we keep marks as just a way to tell pytest which tests to run, and rely on
environment variables to determine what type of minicluster we need to bring up.

Reviewers: yiran, evelynl, adityacb, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3261721