Increase offer timeout value
Summary:
Currently offers are removed from in-memory structure after 30 minutes. If no match happens on that host for 30 minutes, then offer will be removed and since mesos master offer timeout is inifinite, no new offer will come.

Increasing the value to 10 days, such that either a match happens on that host in 10 days or host manager is restarted in this time.

Reviewers: sachins, O521 Project Peloton: Add reviewers, zhixin

Reviewed By: sachins, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3274293