Add support for a separate resource pool for GPU resources in aurora bridge
Summary:
GPU and CPU stateless jobs should not share the same resource pool.
So, allow configuration of a separate resource pool for both kind of jobs.
This commit does not provide the ability to create both resource pools
in the same cluster. It only provides the ability to either submit jobs
to pre-exiting resource pools (of both types) or create only one type of
resource pool.

Test Plan: Unit tests and tested in minicluster.

Reviewers: kevinxu, O521 Project Peloton: Add reviewers

Reviewed By: kevinxu, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3274663