Preempt tasks on mesos-task-id
Summary:
Earlier, resmgr exposed peloton-task-id of the tasks to be preempted. Jobmgr killed these tasks. The issue with this is jobmgr might kill tasks incorrectly if jobmgr processes the KILLED event faster than resmgr and the preemption cycle runs in resmgr before it processes the KILLED event. In such a case, jobmgr might incorrectly kill the newer run of the task (preemption was meant for older run).

This diff addresses the above issue by preempting tasks on mesos-task-id. If the mesos-task-id of the task-to-be-preempted is different from the task in jobmgr then the kill is skipped.

Test Plan: Unit tests

Reviewers: avyas, zhixin, O521 Project Peloton: Add reviewers

Reviewed By: avyas, zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3279417