Minicluster: Add lib function to check if ZK is ready
Summary:
We currently wait 20 seconds every time we bring up a new minicluster to wait
for ZK to come up. This should be closer to 5 seconds and this library function
will allow us to wait until it is ready instead of sleeping for an arbitrary
amount of time.

Also added a util function to find a random free port on the host. This will come in
handy when we decide to isolate miniclusters from one another.

Reviewers: amitbose, adityacb, yiran, O521 Project Peloton: Add reviewers

Reviewed By: amitbose, adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3287583