Add a CLI to rollback a job to a previous configuration
Summary:
If a job needs to be rolled back to a previous configuration, then the user needs
to use the CLI to first fetch the previous configuration, modify the output to
yaml and then call replace job again. During an outage situation, modifying the
output by hand provides a sub-optimal user experience. So, adding a rollback
CLI which will first fetch the previous configuration, and then invoke replace
job to rollback the job to the previous configuration without the user having
to manually modify any configuration files.

Reviewers: zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3326759