Fix flaky test 'TestPreemptionCycle'
Summary:
Before, in `TestPreemptionCycle`, there were 3 expect calls to `PatchTasks`, all identical with respect to expected arguments. An actual call could be matched to any of these PatchTasks call, making the test flaky.

This diff addresses the issue by replacing the 3 expect calls with one call and using `Times()` to specify the number of times we expect it to be called.

Test Plan: Ran test locally multiple times to ensure the flakiness is fixed.

Reviewers: amitbose, O521 Project Peloton: Add reviewers

Reviewed By: amitbose, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T3945261

Differential Revision: https://code.uberinternal.com/D3312431